+++ src/router/generic.js
@@ -1,0 +1,8 @@
+const routes = {
+  path: "/login",
+  name: "generic.login",
+  component: () =>
+    import(/* webpackChunkName: "about" */ "../views/generic/Login.vue")
+};
+
+export default routes;
